Add Azure Event Hubs Emulator container to Azure module - Rename container - Remove Kafka
