+++ modules/azure/build.gradle
@@ -11,5 +11,4 @@ dependencies {
     testImplementation 'com.azure:azure-storage-queue:12.24.0'
     testImplementation 'com.azure:azure-data-tables:12.5.0'
     testImplementation 'com.azure:azure-messaging-eventhubs:5.19.2'
-    testImplementation 'org.apache.kafka:kafka-clients:3.8.0'
 }
